+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/experiments/InAndOuttakeTesting.java
@@ -37,7 +37,7 @@ public class InAndOuttakeTesting extends LinearOpMode {
     private double kI = 0.0;
     private double kD = 0.0;
 
-    private double targetRPM = 5000;
+    private int targetRPM = 5000;
     private double testPower = 0.4;
 
     // Menu Navigation
